+++ misc/dbt-materialize/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Support unmanaged clusters in `deploy_init`. The deployment cluster
+  is now created by cloning the production cluster's replicas (including
+  `SIZE` and `AVAILABILITY ZONE`).
+
 ## 1.9.9 - 2026-05-18
 
 * Add support for the `partition_by` configuration in materialized views,